fixed some broken specs

+++ app/models/taxon.rb
@@ -877,14 +877,14 @@ class Taxon < ActiveRecord::Base
   
   def rank_level_must_be_finer_than_parent
     return if parent.nil?
-    if (parent.rank_level.nil? || rank_level.nil?) || (parent.rank_level.to_i <= rank_level.to_i)
+    if (parent.rank_level.nil? || rank_level.nil?) || (parent.rank_level.to_f <= rank_level.to_f)
       errors.add(self.name, "rank level must be finer than parent")
     end
   end
   
   def rank_level_must_be_coarser_than_children
     return if new_record?
-    if (children.any?{ |e| e.rank_level.nil? }  || rank_level.nil?) || (children.any?{ |e| e.rank_level.to_i >= rank_level.to_i })
+    if (children.any?{ |e| e.rank_level.nil? }  || rank_level.nil?) || (children.any?{ |e| e.rank_level.to_f >= rank_level.to_f })
       errors.add(self.name, "rank level must be coarser than children")
     end
   end
